requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-dotenv